fix(profile): compute the required sign field for visitor-history requests
Root-caused live via TRACE-level upstream HTTP capture: the real HelloTalk
API was rejecting our /user_profile_visitor/v2/my_history request with
{"code":400,"msg":"no data currently"} — a deliberately misleading
message that reads like 'zero visitors' but is actually 'unauthenticated
request'. Our outbound body omitted the sign field entirely (previously
hardcoded to null in ProfileController.visitors()).

Reverse-engineered the exact algorithm from re_output smali (agent-traced,
cross-checked against two independent call sites — my_history and the
sibling v3/visit endpoint, byte-identical scheme in both):
    sign = MD5(jid + jid + client_ts).toLowerCase()
where jid is the account's own numeric uid (not the target user's) and
client_ts is the corrected epoch-ms timestamp, sent as-is in both the
client_ts and update_ts fields of the same request.

- Md5Util.visitorHistorySign(jid, clientTs) — new method, full smali
  citation in the doc comment (bq0/c.smali:1007-1049, bq0/g.smali:190-270).
- ProfileController: injects ObjectMapper to derive callerUserId via
  UidExtractor (same pattern as ImSendController/ImEventSource); visitors()
  now computes and sends a real sign, and sends update_ts as the same
  client_ts value instead of a hardcoded 0.
- VisitorHistoryRequest.updateTs: int -> long. The smali sends a full
  epoch-ms timestamp in this field; an int would silently overflow (current
  epoch-ms is ~1.78 trillion, past Integer.MAX_VALUE).

Verified live: the sign now appears correctly in the outbound request and
independently re-verified byte-for-byte against a standalone MD5
computation. Upstream still returns "no data currently" for this specific
account/moment — either the account genuinely has zero pending visitor
records right now, or this isn't the same account being checked in the
live app (open question from earlier in this engagement, still
unconfirmed). Either way, the request is now correctly authenticated per
the real client's own signing scheme, which it provably was not before.

+++ src/main/java/com/jilali/user/dto/VisitorHistoryRequest.java
@@ -11,6 +11,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * camelCase keys the real upstream silently ignores, confirmed against a live captured
  * request body: {@code {"device_type","client_ts","index","device_id","sign","client_ver",
  * "update_ts","client_os"}}.
+ *
+ * <p>{@code updateTs} is {@code long}, not {@code int} — the smali ({@code bq0/c.smali:958-1055})
+ * sends the same corrected epoch-ms {@code client_ts} value in both fields. An {@code int} here
+ * would silently overflow/wrap a real millisecond timestamp (current epoch-ms is already ~1.78
+ * trillion, well past {@code Integer.MAX_VALUE}).
  */
 @Serdeable
 public record VisitorHistoryRequest(
@@ -20,6 +25,6 @@ public record VisitorHistoryRequest(
     @JsonProperty("device_id") @Nullable String deviceId,
     @Nullable String sign,
     @JsonProperty("client_ver") @Nullable String clientVer,
-    @JsonProperty("update_ts") int updateTs,
+    @JsonProperty("update_ts") long updateTs,
     @JsonProperty("client_os") int clientOs
 ) {}
